+++ src/store/store.js
@@ -6,3 +6,5 @@ export const entryStore = writable(10);
 export const sortDirStore = writable("ascending");
 export const sortByStore = writable("rank");
 export const dataLoading = writable(true);
+export const pageStore = writable(1);
+export const pageCountStore = writable(entryStore);